pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-code-reviewer-py"
version = "0.1.0"
authors = [
  { name="anywaifu", email="anywaifu@protonmail.com" },
]
description = "A CLI tool to review git commits using AI"
readme = "README.md"
requires-python = ">=3.13"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Quality Assurance",
    "Environment :: Console",
]
dependencies = [
    "click>=8.0",
    "litellm>=1.69",
    "gitpython>=3.1",
    "pydantic>=1.10",
    "rich>=13.0",
    "python-dotenv>=1.0",
]

[tool.setuptools.packages.find]
include = ["ai_code_reviewer_py*"]
exclude = ["test_reviews*", "tests*"]

[project.scripts]
ai-code-reviewer = "ai_code_reviewer_py.cli:cli"

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-mock>=3.0", "mypy>=1.0", "pytest-asyncio>=1.0.0"]